+++ WpBehavior/inc/CSensorSmell.h
@@ -94,7 +94,7 @@ public:
 //---------------------------------------------------------------------------
 public:
   const std::vector<sensorValType>& UpdateAndGetStimulationTable();
-  string GetSubCaptorLabel(int index);
+  string GetSubCaptorLabel(size_t index);
 
 };
