+++ WpEnvironment/inc/CScenarioPlayer.h
@@ -98,7 +98,7 @@ public:
   bool IsScenarioFileOpened();
 
   static bool ExecuteCmd(CBiotop* pBiotop, string commandString, string biotopFilesPath, int &customVar1, int &customVar2, CommandHandler_t* pCustomCmdList=NULL, int customCmdListSize=0);
-  static string GetParamFromString(string commandParam,int paramIndex);
+  static string GetParamFromString(string commandParam, size_t paramIndex);
   static string GetHelpCmdString(int index);
 
   static bool CmdLoadBiotop          (CBiotop* pBiotop, string path, string commandParam, int* pSuccessScore, int* pTotalScore);
